Fix a bug the game tries to reload when the last tower is destroyed
The game must defer initialization of game state after the constructor
was finished because the calling logic can assign event handlers for
object creation.  The game has assumed it's not yet initialized if
towers array is empty, but that case can also happen when the enemy
shoots out the last tower.  So we should properly have a dedicated
flag to indicate that the game requires initialization on the first
invocation of update().