Prevent towers from shooting after killing all visible enemies
The towers had reference to dead enemies even if the target's health
was below 0.  We had searched list of enemies every frame to update the
target, so this issue had not been exposed until enemy waves are
implemented.

JavaScript, as a garbage collection language, keeps the enemy alive
if the towers refer to one.  So clearing the reference also helps to
reduse memory usage.