Enable stage level selection mechanism
Instead of endless fight against enemies, the player can select from
list of stages with finite playing time and varying difficulties.
This way, the player can choose low difficulties while his turrets
are weak, in order to gain XP and money.

Current stage's progress is shown as a bar at the bottom of the game
window.  There are currently 10 waves in a stage.

Auto save feature is no longer invoked periodically.  Instead, it
is called when a stage is finished.  So the player should be careful
not to challenge too high difficulties.